refactor: state dropdown location
