+++ web/components/project/index.ts
@@ -7,7 +7,6 @@ export * from "./single-project-card";
 export * from "./single-sidebar-project";
 export * from "./confirm-project-leave-modal";
 export * from "./member-select";
-export * from "./state-select";
 export * from "./members-select";
 export * from "./label-select";
 export * from "./priority-select";
