+++ web/components/states/index.ts
@@ -2,3 +2,4 @@ export * from "./create-update-state-inline";
 export * from "./create-state-modal";
 export * from "./delete-state-modal";
 export * from "./single-state";
+export * from "./state-select";
